Create 1
This code creates a simple to-do list application that allows users to add tasks and delete tasks from a list. The tasks are stored in an array and the array is displayed in a table. The application uses a form with a text input and a submit button to add tasks and another form with a hidden input and a submit button to delete tasks. The array_push function is used to add a task to the array and the unset function is used to delete a task from the array. The array_values function is used to reset the indices of the array after a task has been deleted.